Rewrite callback system
So far, callbacks were managed by registering the callback receiver
in the Lua interpreter and then storing the handle along the entity that
will receive the calls. The script where the entity is declared is
assumed to be the script where callbacks will be invoked.

This previous design prevents the separation of entities in header and
body. As callbacks are received in the same Lua interpreter where entity
was declared. It leads to import cycles when a callbacks of two rooms
have to reference each other.

The solution consists on registering each individual callback into the
user value of the owner entity. This means the callback is not
associated to the whole entity and the script where it is declared, but
to the actual script where the callback is declared. This breaks the
limitations of the previous design, allowing callbacks to be registered
in any script, and therefore separating scripts into header and bodies.