Take entity custom getters from user value
The previous change introduced another problem. As now the callbacks
can be declared in scripts other than those where the entity is
declared, those callbacks might attempt to access a field injected by
the entity constructor. That is the case of rooms, which inject objects
and walkboxes as fields.

To prevent it, this change do not inject the fields but allow an entity
to implement the ScriptCustomGetter interface. If so, the default
__index Lua metamethod will get the custom field from the user value.