Prevent deadlocks caused by callbacks
If a room callback shows another room, and the exit callback of that
very room is invoked, it will have an undefined behavior. As the same
script that was waiting for the show room will be used to invoke the
exit function.

This change adds a mutex to the Lua interpreter that is acquired when
code is being executed and released during waits (for futures and
sleeps).